chore(mcp): rename hog function and template
removing the hog from the MCP tools/descriptions altogether
